Multiple destinations
Animal now has a choice of destinations and can pick a new one each time the behaviour changes. Some issue with the options still as we get a few at 0,0. Next step would be to make it dynamic and a varying number of destinations per behaviour mode. Currently fixed to 3 only and that needs specifiying in nDes. Not ideal, but appears to work. Movement pathways appear between destinations that is awesome.