Dynamic foraging
Dynamic foraging added, but has highlighted a major issue with the sample options function. If given all the same value the sample options function appears to return a choice beyond the total umber of choices. Appears to be recreatable in the R wrapped version also. Have added a very rudimentary error capture for this in the sims, but only catches last instance so might not be helpful in IDing future issues, but least has highlighted the issue with sample_options for now. Break for lunch.