Forage and shelter sites attraction
Both discrete predefined sites and the dynamic foraging sites appear to be working and stable. Still some balancing needed with the behavioural shifts. Had to drop the normalisation in sample_options as we need the negative values to prevent resting during active hours. But the sample_options works without the normalisation now it can handle all similar values.