Integrated avoidance point loop
So far it seems to work and is stable. The impact of the attraction points and the avoidance points needs to be user defined at the simulation stage, but that needs some extra inputs added. Might need a system fro defining if the effect is sqrt/pow. Avoidance point weighting needs careful explaining in documentation: each moveOption has a cumulative distance from all avoidance points calculated, then the the nopt number of cum.dists are standardised.