Changed the additional cycles
Now additional cycles are entirely derived from the inputs and no longer limited to two. The number of cycles that be input is unlimited (although balancing will become harder). Seems to work and is stable. Expanded and made saleable the plotting code for review of the behaviour cycles too. Might run into issues with scales and too many cycles.