Raster building and spinning stopped
We now have a more reasoned raster generation process, and the shelter sites are draw from the environment. We have updated the visuals for the rasters too. Each species now has a unique suite of layers to work with that should be being fed into the sim, but should double check they are being used. 
Also fixed the angle spinning bug, now the angle update is corrected to between -180 and 180 rather than between -360 and 360. Looks good and figures updated to reflect that.