Started justifying inputs and rescale fix
Started adding in more realistic values for the species. But discovered a dire need to detach the matrix units from the step length units. Have solved this by adding a rescale variable that converts the step lengths to the matrix units after they have been drawn from the gamma distributions. The returned step lengths in the results are matching the matrix units, so we now return the rescale factor alongside for easy back transformation.