Plots now take in the rescale for the output review of step lengths.
This is fixed in the vignette. No change needed in mapping. Behaviour mode for badger need adjusting for sure. The total distance travelled from the papers seems to be mismatched with the speed it can go in a minute. Might need to reconsider how we calibrate that for something realistic. Maybe stronger or weaker attraction.