Massive commit changing the structure of the vignettes
Main MS vignette has been updated with new references and moved to notebook. Notebook is a new location that is ignored when building. The vignette once complete should be placed in the vignettes folder that will be included in the build. There is an asis file that should aid this. This new notebook layout also helps with the externally generated figures, ie no longer need an inst folder. The notebook is kept in version control but doesn't get included in the package people download. Also moved a bunch of testing code that was messy to the notebook as we don't need that for building.