Changed sample_options function
Now we are using the Rcpp Sample function, but it requires rebalancing. Also it does not like negative values so we need a solution for creating no-go areas in the landscape, perhaps an exception where sample cannot return an index linked to -99? That will run into issues when the animal doesn't give itself any choices except for -99.... but that shouldn't happen because the current location is always an option. So remain still and try again essentially.